feat: remove route groups in admin folder
