+++ frontend/src/app/admins/[id]/admin-list/[adminId]/page.tsx
