+++ frontend/src/app/admins/[id]/customer-list/[customerId]/page.module.scss
