+++ frontend/src/app/admins/[id]/customer-list/[customerId]/page.tsx
