+++ frontend/src/app/admins/[id]/error.tsx
