+++ frontend/src/app/admins/[id]/event-list/[eventId]/page.tsx
