+++ frontend/src/app/admins/[id]/instructor-list/[instructorId]/page.tsx
