+++ frontend/src/app/admins/[id]/layout.module.scss
