+++ frontend/src/app/admins/[id]/layout.tsx
