+++ frontend/src/app/admins/[id]/loading.tsx
