+++ frontend/src/app/admins/[id]/plan-list/[planId]/page.tsx
