+++ frontend/src/app/admins/[id]/subscription-list/[subscriptionId]/page.tsx
