+++ frontend/src/app/admins/login/page.module.scss
@@ -1,0 +1,48 @@
+@import "../../variables.module.scss";
+
+.outsideContainer {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  background-color: $white-background;
+}
+
+.container {
+  padding: 70px 30px 30px;
+  border: 1px solid $gray-background;
+  border-radius: 0.5rem;
+  box-shadow: 2px 4px 6px rgba(0, 0, 0, 0.3);
+  position: relative;
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  max-width: 350px;
+  width: 100%;
+
+  h2 {
+    margin-bottom: 0.3rem;
+  }
+}
+
+.logo {
+  position: absolute;
+  top: -1rem;
+  left: 1rem;
+}
+
+.buttonWrapper {
+  display: flex;
+  justify-content: center;
+  margin: 1rem;
+}
+
+.icon {
+  width: 1.2rem;
+  height: 1.2rem;
+}
+
+.form {
+  width: 100%;
+}
